Organize all greenplum files into a greenplum directory.
