+++ contrib/pg_upgrade/Makefile
@@ -11,11 +11,12 @@ OBJS = check.o controldata.o dump.o exec.o file.o function.o info.o \
        tablespace.o util.o version.o version_old_8_3.o $(WIN32RES)
 
 # Source files specific to Greenplum
-OBJS += aotable.o gpdb4_heap_convert.o version_gp.o \
-        check_gp.o file_gp.o reporting.o aomd_filehandler.o \
-        old_tablespace_file_contents.o old_tablespace_file_parser.o \
-        tablespace_gp.o old_tablespace_file_parser_observer.o \
-        info_gp.o
+OBJS += aomd_filehandler.o
+OBJS += greenplum/aotable.o greenplum/gpdb4_heap_convert.o greenplum/version_gp.o \
+        greenplum/check_gp.o greenplum/file_gp.o greenplum/reporting.o \
+        greenplum/old_tablespace_file_contents.o greenplum/old_tablespace_file_parser.o \
+        greenplum/tablespace_gp.o greenplum/old_tablespace_file_parser_observer.o \
+        greenplum/info_gp.o
 
 PG_CPPFLAGS  = -DFRONTEND -DDLSUFFIX=\"$(DLSUFFIX)\" -I$(srcdir) -I$(libpq_srcdir)
 PG_LIBS = $(libpq_pgport)
