+++ contrib/pg_upgrade/greenplum/aotable.c
