+++ contrib/pg_upgrade/greenplum/reporting.c
