+++ contrib/pg_upgrade/tablespace.c
@@ -12,7 +12,7 @@
 #include "pg_upgrade.h"
 
 #include <sys/types.h>
-#include "old_tablespace_file_contents.h"
+#include "greenplum/old_tablespace_file_contents.h"
 
 static void get_tablespace_paths(void);
 static void set_tablespace_directory_suffix(ClusterInfo *cluster);
