+++ contrib/pg_upgrade/test/integration/tablespace_gp_test.c
@@ -12,8 +12,8 @@
  * Production dependencies
  */
 #include "pg_upgrade.h"
-#include "old_tablespace_file_parser_observer.h"
-#include "old_tablespace_file_contents.h"
+#include "greenplum/old_tablespace_file_parser_observer.h"
+#include "greenplum/old_tablespace_file_contents.h"
 
 /*
  * Test dependencies
