+++ contrib/pg_upgrade/test/unit/check_gp_test.c
@@ -1,6 +1,6 @@
 #include "cmockery_gp.h"
 
-#include "check_gp.h"
+#include "greenplum/check_gp.h"
 #include "pg_upgrade.h"
 
 ClusterInfo old_cluster,
