+++ contrib/pg_upgrade/test/unit/old_tablespace_file_parser_test.c
@@ -1,7 +1,7 @@
 #include "cmockery_gp.h"
 
-#include "old_tablespace_file_parser.h"
-#include "old_tablespace_file_parser_observer.h"
+#include "greenplum/old_tablespace_file_parser.h"
+#include "greenplum/old_tablespace_file_parser_observer.h"
 
 static int _invalid_access_to_row_called_with_row_index;
 static int _invalid_access_to_field_called_with_row_index;
